requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2020.12.5
    # via requests
chardet==3.0.4
    # via requests
click==7.1.2
    # via typer
idna==2.10
    # via requests
numpy==1.19.4
    # via pandas
oauthlib==3.1.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
pandas==1.2.0
    # via -r requirements.in
py-trello==0.17.1
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   pandas
    #   py-trello
pytz==2020.4
    # via
    #   pandas
    #   py-trello
pyyaml==5.3.1
    # via -r requirements.in
requests-oauthlib==1.3.0
    # via
    #   -r requirements.in
    #   py-trello
requests==2.25.0
    # via
    #   py-trello
    #   requests-oauthlib
six==1.15.0
    # via python-dateutil
typer==0.3.2
    # via -r requirements.in
urllib3==1.26.2
    # via requests
xlsxwriter==1.3.7
    # via -r requirements.in
